requirement.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
djangorestframework==3.12.4
idna==2.10
numpy==1.20.3
pandas==1.2.4
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.1
urllib3==1.26.5
